+++ .idea/workspace.xml
@@ -20,7 +20,16 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="bc972702-0380-448d-9299-7724cfb49770" name="Default Changelist" comment="">
+      <change afterPath="$PROJECT_DIR$/.gitignore" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/js/Storage.js" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/js/TodoApp.js" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/js/TodoCount.js" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/js/TodoInput.js" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/js/TodoItem.js" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/js/TodoList.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/index.html" beforeDir="false" afterPath="$PROJECT_DIR$/index.html" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/js/app.js" beforeDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -59,7 +68,19 @@
       <updated>1598965395327</updated>
       <workItem from="1598965396453" duration="1363000" />
       <workItem from="1599119796279" duration="2059000" />
-      <workItem from="1599133720002" duration="3505000" />
+      <workItem from="1599133720002" duration="3844000" />
+      <workItem from="1599219019048" duration="4675000" />
+      <workItem from="1599482242277" duration="5608000" />
+      <workItem from="1599716049725" duration="4993000" />
+      <workItem from="1599808251335" duration="9380000" />
+      <workItem from="1599833883247" duration="959000" />
+      <workItem from="1599972077441" duration="981000" />
+      <workItem from="1599983733920" duration="7663000" />
+      <workItem from="1600070067712" duration="47000" />
+      <workItem from="1600156050146" duration="365000" />
+      <workItem from="1600157907755" duration="4313000" />
+      <workItem from="1600176051070" duration="6084000" />
+      <workItem from="1600211645968" duration="8229000" />
     </task>
     <servers />
   </component>
@@ -77,5 +98,19 @@
       </map>
     </option>
     <option name="oldMeFiltersMigrated" value="true" />
+  </component>
+  <component name="WindowStateProjectService">
+    <state x="93" y="93" width="1350" height="638" key="DiffContextDialog" timestamp="1600215632372">
+      <screen x="0" y="0" width="1536" height="824" />
+    </state>
+    <state x="93" y="93" width="1350" height="638" key="DiffContextDialog/0.0.1536.824@0.0.1536.824" timestamp="1600215632372" />
+    <state width="498" height="446" key="SwitcherDM" timestamp="1600215463084">
+      <screen x="0" y="0" width="1536" height="824" />
+    </state>
+    <state width="498" height="446" key="SwitcherDM/0.0.1536.824@0.0.1536.824" timestamp="1600215463084" />
+    <state x="372" y="22" key="com.intellij.openapi.editor.actions.MultiplePasteAction$ClipboardContentChooser" timestamp="1600215612648">
+      <screen x="0" y="0" width="1536" height="824" />
+    </state>
+    <state x="372" y="22" key="com.intellij.openapi.editor.actions.MultiplePasteAction$ClipboardContentChooser/0.0.1536.824@0.0.1536.824" timestamp="1600215612648" />
   </component>
 </project>
