+++ index.html
@@ -34,6 +34,6 @@
         </div>
       </main>
     </div>
-    <script src="./js/app.js"></script>
+    <script type="module" src="/js/TodoApp.js"></script>
   </body>
 </html>
