+++ js/Storage.js
@@ -1,0 +1,8 @@
+export function getStorage() {
+    return JSON.parse(localStorage.getItem("todoItems")) || [];
+};
+
+export function setStorage(todoItems) {
+    return localStorage.setItem("todoItems", JSON.stringify(todoItems));
+};
+
