+++ js/TodoItem.js
@@ -1,0 +1,9 @@
+function TodoItem(title, nextId){
+    return {
+        id: nextId,
+        title: title,
+        type: "active"
+    }
+};
+
+export default TodoItem;
